requirements.txt
Flask==3.0.0
pandas==2.0.3
python-dotenv==1.0.0
vobject==0.9.6.1
openpyxl==3.1.2
gunicorn==21.2.0
flask-cors==4.0.0